# api/swagger.yaml
definitions:
...
  v4.Category:
    properties:
      archived:
        default: false
        description: Is the category archived?
        example: true
        type: boolean
      budgetId:
        description: ID of the budget the category belongs to
        example: 52d967d3-33f4-4b04-9ba7-772e5ab9d0ce
        type: string
      createdAt:
        description: Time the resource was created
        example: "2022-04-02T19:28:44.491514Z"
        type: string
      deletedAt:
        description: Time the resource was marked as deleted
        example: "2022-04-22T21:01:05.058161Z"
        type: string
      envelopes:
        description: These fields are computed
        items:
          $ref: '#/definitions/v4.Envelope'
        type: array
      id:
        description: UUID for the resource
        example: 65392deb-5e92-4268-b114-297faad6cdce
        type: string
      links:
        $ref: '#/definitions/v4.CategoryLinks'
      name:
        description: Name of the category
        example: Saving
        type: string
      note:
        description: Notes about the category
        example: All envelopes for long-term saving
        type: string
      updatedAt:
        description: Last time the resource was updated
        example: "2022-04-17T20:14:01.048145Z"
        type: string
    type: object
  v4.CategoryCreateResponse:
    properties:
      data:
        description: List of the created Categories or their respective error
        items:
          $ref: '#/definitions/v4.CategoryResponse'
        type: array
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
    type: object
  v4.CategoryEditable:
    properties:
      archived:
        default: false
        description: Is the category archived?
        example: true
        type: boolean
      budgetId:
        description: ID of the budget the category belongs to
        example: 52d967d3-33f4-4b04-9ba7-772e5ab9d0ce
        type: string
      name:
        description: Name of the category
        example: Saving
        type: string
      note:
        description: Notes about the category
        example: All envelopes for long-term saving
        type: string
    type: object
...
  v4.CategoryLinks:
    properties:
      envelopes:
        description: Envelopes for this category
        example: https://example.com/api/v4/envelopes?category=3b1ea324-d438-4419-882a-2fc91d71772f
        type: string
      self:
        description: The category itself
        example: https://example.com/api/v4/categories/3b1ea324-d438-4419-882a-2fc91d71772f
        type: string
    type: object
  v4.CategoryListResponse:
    properties:
      data:
        description: List of Categories
        items:
          $ref: '#/definitions/v4.Category'
        type: array
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
      pagination:
        allOf:
        - $ref: '#/definitions/v4.Pagination'
        description: Pagination information
    type: object
  v4.CategoryResponse:
    properties:
      data:
        allOf:
        - $ref: '#/definitions/v4.Category'
        description: Data for the Category
      error:
        description: The error, if any occurred
        example: the specified resource ID is not a valid UUID
        type: string
    type: object
  v4.Envelope:
    properties:
      archived:
        default: false
        description: Is the envelope archived?
        example: true
        type: boolean
      categoryId:
        description: ID of the category the envelope belongs to
        example: 878c831f-af99-4a71-b3ca-80deb7d793c1
        type: string
      createdAt:
        description: Time the resource was created
        example: "2022-04-02T19:28:44.491514Z"
        type: string
      deletedAt:
        description: Time the resource was marked as deleted
        example: "2022-04-22T21:01:05.058161Z"
        type: string
      id:
        description: UUID for the resource
        example: 65392deb-5e92-4268-b114-297faad6cdce
        type: string
      links:
        allOf:
        - $ref: '#/definitions/v4.EnvelopeLinks'
        description: Links to related resources
      name:
        description: Name of the envelope
        example: Groceries
        type: string
      note:
        description: Notes about the envelope
        example: For stuff bought at supermarkets and drugstores
        type: string
      updatedAt:
        description: Last time the resource was updated
        example: "2022-04-17T20:14:01.048145Z"
        type: string
    type: object
...
  v4.EnvelopeLinks:
    properties:
      month:
        description: The MonthConfig for the envelope
        example: https://example.com/api/v4/envelopes/45b6b5b9-f746-4ae9-b77b-7688b91f8166/YYYY-MM
        type: string
      self:
        description: The envelope itself
        example: https://example.com/api/v4/envelopes/45b6b5b9-f746-4ae9-b77b-7688b91f8166
        type: string
      transactions:
        description: The envelope's transactions
        example: https://example.com/api/v4/transactions?envelope=45b6b5b9-f746-4ae9-b77b-7688b91f8166
        type: string
    type: object
...
  v4.Pagination:
    properties:
      count:
        description: The amount of records returned in this response
        example: 25
        type: integer
      limit:
        description: The maximum amount of resources to return for this request
        example: 25
        type: integer
      offset:
        description: The offset for the first record returned
        example: 50
        type: integer
      total:
        description: The total number of resources matching the query
        example: 827
        type: integer
    type: object
...
  v4.httpError:
    properties:
      error:
        example: An ID specified in the query string was not a valid UUID
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /v4/categories:
    get:
      description: Returns a list of categories
      parameters:
      - description: Filter by name
        in: query
        name: name
        type: string
      - description: Filter by note
        in: query
        name: note
        type: string
      - description: Filter by budget ID
        in: query
        name: budget
        type: string
      - description: Is the category archived?
        in: query
        name: archived
        type: boolean
      - description: Search for this text in name and note
        in: query
        name: search
        type: string
      - description: The offset of the first Category returned. Defaults to 0.
        in: query
        name: offset
        type: integer
      - description: Maximum number of Categories to return. Defaults to 50.
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.CategoryListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.CategoryListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.CategoryListResponse'
      summary: Get categories
      tags:
      - Categories
    options:
      description: Returns an empty response with the HTTP Header "allow" set to the
        allowed HTTP verbs
      responses:
        "204":
          description: No Content
      summary: Allowed HTTP verbs
      tags:
      - Categories
    post:
      description: Creates a new category
      parameters:
      - description: Categories
        in: body
        name: categories
        required: true
        schema:
          items:
            $ref: '#/definitions/v4.CategoryEditable'
          type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v4.CategoryCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.CategoryCreateResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.CategoryCreateResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.CategoryCreateResponse'
      summary: Create category
      tags:
      - Categories
  /v4/categories/{id}:
    delete:
      description: Deletes a category
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.httpError'
      summary: Delete category
      tags:
      - Categories
    get:
      description: Returns a specific category
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.CategoryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.CategoryResponse'
      summary: Get category
      tags:
      - Categories
    options:
      description: Returns an empty response with the HTTP Header "allow" set to the
        allowed HTTP verbs
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.httpError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.httpError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.httpError'
      summary: Allowed HTTP verbs
      tags:
      - Categories
    patch:
      consumes:
      - application/json
      description: Update an existing category. Only values to be updated need to
        be specified.
      parameters:
      - description: ID formatted as string
        in: path
        name: id
        required: true
        type: string
      - description: Category
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/v4.CategoryEditable'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v4.CategoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v4.CategoryResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v4.CategoryResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v4.CategoryResponse'
      summary: Update category
      tags:
      - Categories
...
swagger: "2.0"